+ after 11:00

+++ src/main/java/ru/javawebinar/lunchvoting/repository/VoteRepository.java
@@ -2,20 +2,25 @@ package ru.javawebinar.lunchvoting.repository;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import ru.javawebinar.lunchvoting.model.Menu;
 import ru.javawebinar.lunchvoting.model.User;
 import ru.javawebinar.lunchvoting.model.Vote;
+import ru.javawebinar.lunchvoting.util.exception.IllegalRequestDataException;
 import ru.javawebinar.lunchvoting.util.exception.NotFoundException;
 
 import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
 public class VoteRepository {
+    public static LocalTime TIME_CHANGE_VOTE = LocalTime.of(11, 00);
+    public static LocalDate DATE_NOW_FOR_TEST_UPDATE = LocalDate.of(2020, 05, 01);
+    public static LocalTime TIME_NOW_FOR_TEST_UPDATE = LocalTime.of(11, 00);
+
     protected final Logger log = LoggerFactory.getLogger(getClass());
 
     private final CrudVoteRepository crudVoteRepository;
@@ -78,6 +83,23 @@ public class VoteRepository {
                 .orElseThrow(() -> new NotFoundException("Not found menu with " + menuId));
         Vote voteOld = Optional.of(getByDateLunch(newMenu.getDateMenu(), userId))
                 .orElseThrow(() -> new NotFoundException("Not found vote for update with " + menuId));
+
+        // for testing  // for testing  //  for testing
+        LocalDate testLocalDateNow = DATE_NOW_FOR_TEST_UPDATE;// date for testing
+        LocalTime testLocalTimeNow = TIME_NOW_FOR_TEST_UPDATE;// time for testing
+
+        // if date vote < date now then error
+        if (voteOld.getDateLunch().isBefore(testLocalDateNow)) {
+            throw new IllegalRequestDataException("date is Before now() menuId=" + menuId);
+        }
+        // check time after 11:00
+        if (voteOld.getDateLunch().isEqual(testLocalDateNow)) {
+            if (testLocalTimeNow.isAfter(TIME_CHANGE_VOTE)) {
+                throw new IllegalRequestDataException("time is after 11:00 for update menuId " + menuId);
+            }
+        }
+
+        // if date vote > date now  or time < 11:00
         crudVoteRepository.deleteByIdUserId(voteOld.getId(), userId);
         save(menuId, userId);
     }
